random adjustments to testtype mega crud
