+++ app/Models/TestType.php
@@ -23,19 +23,9 @@ class TestType extends Model
         return $this->hasMany('App\Models\Test');
     }
 
-    public function testtypecategory()
+    public function testTypeCategory()
     {
         return $this->hasOne('App\Models\TestTypeCategory', 'id', 'test_type_category_id');
-    }
-
-    public function specimenType()
-    {
-        return $this->belongsToMany('App\Models\SpecimenType');
-    }
-
-    public function specimenTypeTestType()
-    {
-        return $this->hasMany('App\Models\specimenTypeTestType');
     }
 
     public function specimenTypes()
